+++ provider/pkg/gen/schema.go
@@ -507,7 +507,6 @@ Use the navigation below to see detailed documentation for each of the supported
 		"requires": map[string]string{
 			"pulumi":   ">=3.25.0,<4.0.0",
 			"requests": ">=2.21,<3.0",
-			"pyyaml":   ">=5.3.1",
 		},
 		"moduleNameOverrides": modToPkg,
 		"compatibility":       kubernetes20,