+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 ## Unreleased
 
+- [sdk/python] Drop unused pyyaml dependency (https://github.com/pulumi/pulumi-kubernetes/pull/2502)
+
 ## 4.0.1 (July 19, 2023)
 
 - Gracefully handle undefined resource schemes (https://github.com/pulumi/pulumi-kubernetes/pull/2504)
 - Fix diff for CRD .spec.preserveUnknownFields (https://github.com/pulumi/pulumi-kubernetes/pull/2506)
+
 
 ## 4.0.0 (July 19, 2023)